Refactor extension API handling by removing `extensionApiHandler` and integrating its logic directly into the background channel agent. This change streamlines message processing and enhances clarity by consolidating API request handling within the `runInpageBackgroundChannelBackgroundAgent` function.
