+++ clients/extension/src/background/index.ts
@@ -1,9 +1,16 @@
 import { initializeMessenger } from '@clients/extension/src/messengers/initializeMessenger'
+import { attempt } from '@lib/utils/attempt'
+import { matchRecordUnion } from '@lib/utils/matchRecordUnion'
+import { getRecordUnionKey } from '@lib/utils/record/union/getRecordUnionKey'
+import { getRecordUnionValue } from '@lib/utils/record/union/getRecordUnionValue'
+import { Result } from '@lib/utils/types/Result'
 
 import { ExtensionApiMessage } from '../api'
-import { extensionApiHandler } from '../api/handlers/extensionApiHandler'
 import { runInpageBackgroundChannelBackgroundAgent } from '../channels/inpageBackground/background'
+import { callPopupApiFromBackground } from '../popup/api/call/resolvers/background'
 import { MessageKey } from '../utils/constants'
+import { backgroundApi } from './api'
+import { BackgroundApiMethodName } from './api/interface'
 import { dispatchMessage } from './dispatcher/messageDispatcher'
 import { keepAliveHandler } from './handlers/keepAliveHandler'
 if (!navigator.userAgent.toLowerCase().includes('firefox')) {
@@ -40,13 +47,27 @@ inpageMessenger.reply<{ type: MessageKey; message: any }, unknown>(
 )
 keepAliveHandler()
 
-runInpageBackgroundChannelBackgroundAgent({
+runInpageBackgroundChannelBackgroundAgent<ExtensionApiMessage, Result>({
   handleRequest: async ({ message, context, reply }) => {
-    const extensionMessage = message as ExtensionApiMessage
-    await extensionApiHandler({
-      message: extensionMessage,
-      context,
-      reply,
-    })
+    const result = await attempt(
+      matchRecordUnion<ExtensionApiMessage, Promise<unknown>>(message, {
+        background: backgroundMessage => {
+          const methodName = getRecordUnionKey(backgroundMessage.call)
+          const input = getRecordUnionValue(backgroundMessage.call)
+
+          const resolver = backgroundApi[methodName as BackgroundApiMethodName]
+
+          return resolver({ input, context })
+        },
+        popup: popupMessage => {
+          return callPopupApiFromBackground({
+            call: popupMessage.call,
+            options: popupMessage.options,
+          })
+        },
+      })
+    )
+
+    reply(result)
   },
 })
